requirements.txt
Flask
gunicorn
mysql-connector-python
speechbrain
SpeechRecognition
pydub
torchaudio
torch
soundfile